Josef Changes

+++ src/icTrento.h
@@ -31,13 +31,13 @@ private:
  int nsmoothx;  // smoothly distribute to +- this many cells
  int nsmoothy;
  int nsmoothz;
- void makeSmoothTable(int npart);
+ void makeSmoothTable(double npart);
 
 public:
  IcTrento(Fluid *f, const char *filename, double tau0, const char* setup);
  ~IcTrento();
  double interpolateGrid(double x, double y);
  void setIC(Fluid *f, EoS *eos);
- double setNormalization(int npart);
- double setBaryonNorm(int npart);
+ double setNormalization(double npart);
+ double setBaryonNorm(double npart);
 };
